Correct writeLongToBuffer fix
I did not notice the difference in parameters, data is an unsigned
32-bit integer in one inline function and a *signed* 32-bit integer in
the other. In the two places it was used the data argument was a long.
These calls both casted speed to a signed 8-bit integer, then an
unsigned 8-bit integer, so I added similar casts to encoder.

+++ BaseI2CDevice.h
@@ -83,7 +83,7 @@ inline void writeLongToBuffer(uint8_t* buf, uint32_t data)
 	buf[3] = (data >> 24) & 0xFF;
 }
 
-inline void writeLongToBuffer(uint8_t* buf, long data)
+inline void writeLongToBuffer(uint8_t* buf, int32_t data)
 {
 	writeLongToBuffer(buf, (uint32_t)data);
 }
